Re-drew the IDE page
