+++ client/src/App.js
@@ -3,11 +3,14 @@ import {BrowserRouter, Routes, Route, Link} from 'react-router-dom'
 import Error from './Pages/Error';
 import Register from './Pages/Register';
 import ProtectedRoute from "./Pages/ProtectedRoute";
-import { IDEPage, Profile, SharedLayout } from "./Pages/Dashboard.js";
-import Problem from "./Pages/Dashboard.js/Problem";
+import SharedLayout from "./Pages/Dashboard/SharedLayout";
+import IdePage from "./Pages/IdePage";
+import Profile from "./Pages/Profile";
+import ProblemList from "./Pages/ProblemList";
 import SumTwo from "./Component/SumTwo.js";
 import Palindrome from "./Component/Palindrome.js";
 import AddTwo from "./Component/AddTwo.js";
+import './App.css';
 
 
 function App() {
@@ -17,21 +20,21 @@ function App() {
     <BrowserRouter>
       <Routes>
       <Route path="/" element = {<ProtectedRoute><SharedLayout></SharedLayout></ProtectedRoute>}>
-          <Route index element = {<IDEPage></IDEPage>}></Route>
-          <Route path = "ide" element = {<IDEPage></IDEPage>}>
+          <Route index element = {<IdePage/>}></Route>
+          <Route path = "ide" element = {<IdePage/>}>
             <Route path = "sum-two" element = {<SumTwo/>}/>
             <Route path = "add-two" element = {<AddTwo/>}/>
             <Route path = "palindrome" element = {<Palindrome/>}/>
           </Route>
-          <Route path = "profile" element = {<Profile></Profile>}></Route>
-          <Route path = "question" element = {<Problem/>}></Route>
+          <Route path = "profile" element = {<Profile/>}></Route>
+          <Route path = "question" element = {<ProblemList/>}></Route>
         </Route>
         <Route path="/register" element={<Register/>} />
         <Route path="/landing" element={<Landing />} />
         <Route path="*" element={<Error/>}/>
       </Routes>
     </BrowserRouter>
-    
+
   );
 }
