--- client/src/Pages/Dashboard.js/index.js
@@ -1,11 +1,0 @@
-import SharedLayout from "./SharedLayout";
-import IDEPage from "./IDEPage";
-import Profile from "./Profile";
-
-
-
-export {
-    IDEPage,
-    Profile,
-    SharedLayout,
-}
